# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAphoto API
  description: |
    This is the API for WASAphoto made by 1982048
  version: 1.0.0
...
paths:
...
  /user/{username}/setusername:
    parameters:
      - $ref: "#/components/parameters/username"
    put: 
      security:
      - bearerAuth : []
      tags: ["user"]
      summary: Set username
      description: |-
        Set a new username
      operationId: setMyUserName
      requestBody:
        description: User username
        content: 
          application/json:
            schema:
              $ref: "#/components/schemas/user"
        required: true
      responses:
        "200":
          description: Username set successfully.
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/user"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /user/{username}/stream:
    parameters:
      - $ref: "#/components/parameters/username"
    get:
      security:
      - bearerAuth : []
      tags: ["user"]
      summary: Return stream
      description: |-
        Return a list of photos for the user stream.
      operationId: getMyStream
      responses:
        "200": 
          description: User stream retrived successfully.
          content:
            application/json:
              schema: 
                type: array
                minItems: 0
                description: image array schema.
                maxItems: 999
                items: { $ref: "#/components/schemas/stream"}
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /users/{username}/profile:
    parameters: 
      - $ref: "#/components/parameters/username"
    get:
      security:
      - bearerAuth : []
      tags: ["user"]
      summary: Return profile
      operationId: getUserProfile
      description: |-
        Return an user profile.
      responses:
        "200": 
          description: User profile retrived.
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/profile"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /users/{username}/photo/{photoId}:
    parameters:
      - $ref : "#/components/parameters/username"
      - $ref : "#/components/parameters/photoID"
    put: 
      security:
      - bearerAuth : []
      tags: ["photo"]
      summary: Upload photo
      description: |-
        Upload a photo.
      operationId: uploadPhoto 
      requestBody:
        description: photo
        required: true
        content:
          image/*:
            schema:
              description: Image schema.
              type: object
              properties:
                 file: 
                  description: file
                  type: string
                  format: binary
                  minLength: 0
                  maxLength: 9999   
      responses:
        "201":
          description: Photo uploaded successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/photo"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    delete:
      security:
      - bearerAuth : []
      tags: ["photo"]
      operationId: deletePhoto
      summary: Remove photo
      description: |-
       Delete a photo.
      responses:
        "204": { description: Like deleted successfully. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" } 
  /users/{username}/photo:
    parameters:
      - $ref : "#/components/parameters/username"
    get:
      security:
      - bearerAuth : []   
      tags: ["photo"]
      summary: Return photos
      operationId: getUserPhotos
      description: |-
        Return user photos.
      responses:
        "200": 
          description: User photos retrived.
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/photosList"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /users/{username}/photo/{photoId}/comment/{commentId}:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/photoID"
      - $ref: "#/components/parameters/commentID"
    put:
      security:
      - bearerAuth : []
      tags: ["comment"]
      summary: Add comment
      operationId: commentPhoto
      description: |-
       Add a comment under a specific photo.
      requestBody:
        description: Comment 
        content: 
          application/json:
            schema:
              $ref: "#/components/schemas/comment"
        required: true
      responses:
        "201":
          description:  Comment added successfully.
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/commentsList"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    delete:
     security:
     - bearerAuth : []
     tags: ["comment"]
     operationId: uncommentPhoto
     summary: Remove comment
     description: |-
       Deletete a comment under a specific photo.
     responses:
       "204": { description: Comment deleted successfully. }
       "400": { $ref: "#/components/responses/BadRequest" }
       "401": { $ref: "#/components/responses/UnauthorizedError" }
       "500": { $ref: "#/components/responses/InternalServerError" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses: 
    UnauthorizedError:
      description: Access token is missing or invalid
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc).
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs.
  schemas:
    user:
      title: User 
      description: User schema.
      type: object
      properties:
        identifier:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: "#U1234"
          minLength: 3
          maxLength: 16
        username:
          description: User username.
          type: string
          pattern: '^.*?$'
          example: Francescoerrico
          minLength: 3
          maxLength: 16
    profile:
      title: Profile
      description: Profile schema.
      type: object
      properties:
        requestid:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: "#U1234"
          minLength: 3
          maxLength: 16
        identifier:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: "#U1234"
          minLength: 3
          maxLength: 16
        username:
          description: User username.
          type: string
          pattern: '^.*?$'
          example: Francescoerrico
          minLength: 3
          maxLength: 16
        followersCount:
          description: Number of followers.
          type: integer
          example: 0
          minimum: 0
          maximum: 9999
        followingsCount:
          description: Number of followings.
          type: integer
          example: 0
          minimum: 0
          maximum: 9999
        photoCount:
          description: Number of photos.
          type: integer
          example: 0
          minimum: 0
          maximum: 9999
        followStatus:
          description: User follow status.
          type: boolean
          example: false
        banStatus:
          description: User ban status.
          type: boolean
          example: false
        checkIfBanned:
          description: Check if user is banned.
          type: boolean
          example: false
...
    commentsList: 
      title: Comments
      description: Comments schema.
      type: object
      properties:
        requestId:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: "#U1234"
          minLength: 3
          maxLength: 16
        photoId: 
          description:  Unique photo identifier.
          type: string
          pattern: '^.*?$'
          example: "#P1234"
          minLength: 3
          maxLength: 16
        photoOwner:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: "#U1234"
          minLength: 3
          maxLength: 16
        comments: 
          description: list of objects (each object is a comment).
          type: array
          minItems: 0
          maxItems: 99999
          items:
            type: object
            properties:
              id:
                description: Unique comment identifier.
                type: string
                pattern: '^.*?$'
                example: "#C1234"
                minLength: 3
                maxLength: 16
              userId:
                description: Unique user identifier.
                type: string
                pattern: '^.*?$'
                example: "#U5678" 
                minLength: 3
                maxLength: 16
              photoId: 
                description:  Unique photo identifier.
                type: string
                pattern: '^.*?$'
                example: "#P1234"
                minLength: 3
                maxLength: 16
              photoOwner:
                description: Unique user identifier.
                type: string
                pattern: '^.*?$'
                example: "#U1234"
                minLength: 3
                maxLength: 16
              photoOwnerUsername:
                description: User username.
                type: string
                pattern: '^.*?$'
                example: Francescoerrico
                minLength: 3
                maxLength: 16
              username:
                description: User username.
                type: string
                pattern: '^.*?$'
                example: Francescoerrico
                minLength: 3
                maxLength: 16
              content:
                description: Content of the comment.
                type: string
                pattern: '^.*?$'
                example: "Hello World!"
                minLength: 2
                maxLength: 30
...
    photosList: 
      title: Photos
      description: Photos schema.
      type: object
      properties:
        requestId:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: "#U1234" 
          minLength: 3
          maxLength: 16
        identifier:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: "#U1234" 
          minLength: 3
          maxLength: 16
        photos: 
          description: list of objects (each object is a photo)
          type: array
          minItems: 0
          maxItems: 99999
          items:
            type: object
            properties:
              Id: 
                description:  Unique photo identifier.
                type: string
                pattern: '^.*?$'
                example: "#P1234"
                minLength: 3
                maxLength: 16
              userId:
                description: Unique user identifier.
                type: string
                pattern: '^.*?$'
                example: "#U5678" 
                minLength: 3
                maxLength: 16
              file: 
                description: file
                type: string
                format: binary
                minLength: 3
                maxLength: 16
              date:
                description: Registration date.
                type: string
                format: date-time
                example: "2022-11-02T09:00:00Z"
                minLength: 3
                maxLength: 9999
              likeCount:
                description: photo likes count.
                type: integer
                example: 0
                minimum: 0
                maximum: 9999
              commentCount:
                description: photo comments count.
                type: integer
                example: 0
                minimum: 0
                maximum: 9999
              likeStatus:
                description: photo like status.
                type: boolean
                example: true
    stream: 
      title: Stream
      description: Stream schema.
      type: object
      properties:
        identifier:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: "#U1234" 
          minLength: 3
          maxLength: 16
        photos: 
          description: list of objects (each object is a photo)
          type: array
          minItems: 0
          maxItems: 99999
          items:
            type: object
            properties:
              Id: 
                description:  Unique photo identifier.
                type: string
                pattern: '^.*?$'
                example: "#P1234"
                minLength: 3
                maxLength: 16
              userId:
                description: Unique user identifier.
                type: string
                pattern: '^.*?$'
                example: "#U5678" 
                minLength: 3
                maxLength: 16
              username:
                description: User username.
                type: string
                pattern: '^.*?$'
                example: Francescoerrico
                minLength: 3
                maxLength: 16
              file: 
                description: file
                type: string
                format: binary
                minLength: 3
                maxLength: 16
              date:
                description: Registration date.
                type: string
                format: date-time
                example: "2022-11-02T09:00:00Z"
                minLength: 3
                maxLength: 9999
              likeCount:
                description: photo likes count.
                type: integer
                example: 0
                minimum: 0
                maximum: 9999
              commentCount:
                description: photo comments count.
                type: integer
                example: 0
                minimum: 0
                maximum: 9999
              likeStatus:
                description: photo like status.
                type: boolean
                example: true
...
    photo:
      title: Photo
      description: photo schema.
      properties:
        Id: 
          description:  Unique photo identifier.
          type: string
          pattern: '^.*?$'
          example: "#P1234"
          minLength: 3
          maxLength: 16
        userId:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: "#U5678" 
          minLength: 3
          maxLength: 16
        file: 
            description: file
            type: string
            format: binary
            minLength: 1
            maxLength: 9999
        date:
          description: Registration date.
          type: string
          format: date-time
          example: "2022-11-02T09:00:00Z"
          minLength: 3
          maxLength: 9999
        likeCount:
          description: photo likes count.
          type: integer
          example: 0
          minimum: 0
          maximum: 9999
        commentCount:
          description: photo comments count.
          type: integer
          example: 0
          minimum: 0
          maximum: 9999
        likeStatus:
          description: photo like status.
          type: boolean
          example: true      
    comment:
      title: Comment
      type: object
      description: comment schema.
      properties:
        content:
          description: content
          type: string
          example: Hello World!
          pattern: '^.*?$'
          minLength: 1
          maxLength: 40  
  parameters:
    username:
      schema:
          description: username schema.
          type: string
          pattern: '^.*?$'
          example: Francescoerrico
          minLength: 5
          maxLength: 30
      name: username
      in: path
      required: true
      description: "**Username** of a WASAPhoto user."
    photoID:
      schema:
        description: photoID schema.
        type: string
        pattern: '^.*?$'
        example: "#P1234"
        minLength: 3
        maxLength: 16
      name: photoId 
      in: path
      required: true
      description: "**photoID** of a specific photo on a user profile."
    commentID:
      schema:
        description: commentID schema.
        type: string
        pattern: '^.*?$'
        example: "#C1234"
        minLength: 3
        maxLength: 16
      name: commentId
      in: path
      required: true
      description: "**commentID** of a single comment under a photo."
...